# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API's MrSoft News",
        "description": "API's for MrSoft News",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://mrsoft-news.test",
            "description": "Servidor de la API (dinámico: usa APP_URL del entorno actual)"
        }
    ],
    "paths": {
...
        "/api/contact": {
            "post": {
                "tags": [
                    "Contact"
                ],
                "summary": "Send a contact form email",
                "operationId": "bc749962129af50f203b1defb75ada5e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ruc",
                                    "razon_social",
                                    "persona_contacto",
                                    "telefono",
                                    "correo",
                                    "mensaje",
                                    "producto"
                                ],
                                "properties": {
                                    "ruc": {
                                        "type": "string",
                                        "maxLength": 11,
                                        "minLength": 11,
                                        "example": "12345678901"
                                    },
                                    "razon_social": {
                                        "type": "string",
                                        "example": "Mi Empresa SAC"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Av. Principal 123",
                                        "nullable": true
                                    },
                                    "ciudad_pais": {
                                        "type": "string",
                                        "example": "Lima, Perú",
                                        "nullable": true
                                    },
                                    "persona_contacto": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "+51 999 999 999"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "juan@empresa.com"
                                    },
                                    "mensaje": {
                                        "type": "string",
                                        "maxLength": 1000,
                                        "minLength": 50,
                                        "example": "Estoy interesado en su producto y me gustaría recibir más información sobre precios y funcionalidades."
                                    },
                                    "producto": {
                                        "type": "string",
                                        "example": "Gesrest"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Email enviado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Email enviado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/contactByValues": {
            "post": {
                "tags": [
                    "Contact"
                ],
                "summary": "Send a styled email with custom key-value data",
                "operationId": "0073eb5349aa9aaa57bddf6e790d35f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "emails",
                                    "values"
                                ],
                                "properties": {
                                    "primaryColor": {
                                        "type": "string",
                                        "example": "#FF5733",
                                        "nullable": true
                                    },
                                    "secondaryColor": {
                                        "type": "string",
                                        "example": "#33C3FF",
                                        "nullable": true
                                    },
                                    "foreground": {
                                        "type": "string",
                                        "example": "#FFFFFF",
                                        "nullable": true
                                    },
                                    "emails": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "recipient@empresa.com"
                                        }
                                    },
                                    "values": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "key": {
                                                    "type": "string",
                                                    "example": "Nombre"
                                                },
                                                "value": {
                                                    "type": "string",
                                                    "example": "Juan Pérez"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Email enviado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Email enviado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationError": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "The pagination must be an integer."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}